fix(lint): ESLint mechanical tier — 27 errors fixed (#292)
Clears no-unused-vars, no-empty-object-type (shadcn ui/* interfaces to
type aliases), no-unused-expressions (ternary side-effects to if/else),
and react-refresh/only-export-components (extract non-component exports:
useEventMatcher, useGenerationProgress, profileIdsToApi/apiToProfileIds
into dedicated hook/lib files). Verified via tsc -b + vite build.

40 behavioral React Compiler errors remain (setState-in-effect,
preserve-manual-memoization, etc.) — staged separately as teamarrv2-9y67.3
since they need per-component refactoring and Playwright verification.
