+++ frontend/src/components/ui/alert.tsx
@@ -53,4 +53,4 @@ const Alert = React.forwardRef<HTMLDivElement, AlertProps>(
 )
 Alert.displayName = "Alert"
 
-export { Alert, alertVariants }
+export { Alert }
