+++ frontend/src/components/ui/badge.tsx
@@ -38,4 +38,4 @@ function Badge({ className, variant, ...props }: BadgeProps) {
   )
 }
 
-export { Badge, badgeVariants }
+export { Badge }
