+++ frontend/src/pages/Dashboard.tsx
@@ -3,7 +3,8 @@ import { Button } from "@/components/ui/button"
 import { RunHistoryTable } from "@/components/RunHistoryTable"
 import { EpgOutput } from "@/components/EpgOutput"
 import { ManagedChannelsTable } from "@/components/ManagedChannelsTable"
-import { EventMatcherModal, useEventMatcher } from "@/components/EventMatcherModal"
+import { EventMatcherModal } from "@/components/EventMatcherModal"
+import { useEventMatcher } from "@/hooks/useEventMatcher"
 import { StatusStrip } from "@/components/StatusStrip"
 import { useRecentRuns, useStats } from "@/hooks/useEPG"
 import { usePersistentCollapse } from "@/hooks/usePersistentCollapse"
